+++ src/settings/Configuration.js
@@ -31,25 +31,8 @@ class Configuration extends React.Component {
     merge(stripes, data);
 
     // dispatch locale to force the stripes-intl context to re-render,
-    // allowing the suppressIntlErrors setting to take effect
+    // allowing the suppressIntl* settings to take effect
     stripes.setLocale(stripes.locale);
-
-    if ('serviceWorker' in navigator) {
-      navigator.serviceWorker.getRegistration()
-        .then(reg => {
-          const sw = reg.installing || reg.waiting || reg.active;
-          if (sw) {
-            stripes.logger.log('rtr', 'sending LOGGER_CONFIG');
-            sw.postMessage({ source: '@folio/stripes-core', type: 'LOGGER_CONFIG', value: { categories: data.logger.categories } });
-          } else {
-            // eslint-disable-next-line no-console
-            console.error('error sending LOGGER; sw not registered');
-          }
-        });
-    } else {
-      // eslint-disable-next-line no-console
-      console.error('error sending LOGGER; serviceWorker not found in navigator');
-    }
 
     this.forceUpdate();
   }